university-cost-mcp/requirements.txt
# MCP Server
mcp>=0.9.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
playwright>=1.40.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# API Integration
httpx>=0.25.0
aiohttp>=3.9.0

# LLM for extraction
anthropic>=0.7.0
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
schedule>=1.2.0
